feat(backend): add first unit test and confirm test db reset after each test
The first unit test of the crud functions is done a test.db file is created and cleared after each
test run. The testing script still creates "issues.db" dduring the pytest due to the import of app
for the TestClient so we still need to figure that out. The tests DO use only test.db